+++ Kamienica/src/main/java/kamienica/controller/view/PaymentController.java
@@ -19,17 +19,17 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
+import kamienica.conventer.ApartmentIB;
+import kamienica.conventer.MeterEnergyIB;
+import kamienica.conventer.MeterGasIB;
+import kamienica.conventer.MeterWaterIB;
+import kamienica.conventer.ReadingInvoiceIB;
 import kamienica.core.ManagerEnergy;
 import kamienica.core.ManagerGas;
 import kamienica.core.ManagerPayment;
 import kamienica.core.ManagerWater;
 import kamienica.forms.PaymentForm;
 import kamienica.forms.ReadingInvoiceForm;
-import kamienica.initBinder.ApartmentIB;
-import kamienica.initBinder.MeterEnergyIB;
-import kamienica.initBinder.MeterGasIB;
-import kamienica.initBinder.MeterWaterIB;
-import kamienica.initBinder.ReadingInvoiceIB;
 import kamienica.model.Apartment;
 import kamienica.model.Division;
 import kamienica.model.Invoice;
