+++ Kamienica/src/main/java/kamienica/controller/view/ReadingController.java
@@ -22,13 +22,13 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
+import kamienica.conventer.ApartmentIB;
+import kamienica.conventer.MeterEnergyIB;
+import kamienica.conventer.MeterGasIB;
+import kamienica.conventer.MeterWaterIB;
 import kamienica.forms.ReadingEnergyForm;
 import kamienica.forms.ReadingGasForm;
 import kamienica.forms.ReadingWaterForm;
-import kamienica.initBinder.ApartmentIB;
-import kamienica.initBinder.MeterEnergyIB;
-import kamienica.initBinder.MeterGasIB;
-import kamienica.initBinder.MeterWaterIB;
 import kamienica.model.Apartment;
 import kamienica.model.MeterEnergy;
 import kamienica.model.MeterGas;
