+++ Kamienica/src/main/java/kamienica/conventer/MeterEnergyIB.java
@@ -1,21 +1,21 @@
-package kamienica.initBinder;
+package kamienica.conventer;
 
 import java.beans.PropertyEditorSupport;
 
-import kamienica.model.MeterGas;
+import kamienica.model.MeterEnergy;
 import kamienica.service.MeterService;
 
-public class MeterGasIB extends PropertyEditorSupport {
+public class MeterEnergyIB extends PropertyEditorSupport {
 
 	private final MeterService licznikService;
 	
-	public MeterGasIB(MeterService licznikService) {
+	public MeterEnergyIB(MeterService licznikService) {
 		this.licznikService= licznikService;
 			}
 	
 	@Override
 	public void setAsText(String text) throws IllegalArgumentException {
-		MeterGas licznik = licznikService.getGasByID(Integer.valueOf(text));
+		MeterEnergy licznik = licznikService.getEnergyByID(Integer.valueOf(text));
         setValue(licznik);
 }
 }
